Address composer review findings
