+++ packages/editor/src/send-action.tsx
@@ -1,3 +1,5 @@
+import { ArrowUpIcon } from "@phosphor-icons/react";
+
 export function SendAction(
 	{ busy, label, disabled, onClick }: {
 		busy?: boolean;
@@ -16,9 +18,7 @@ export function SendAction(
 			title={label}
 			type="button"
 		>
-			<svg aria-hidden="true" className="send-action-icon" viewBox="0 0 18 18">
-				<path d="M9 16V3M13.25 7.75 9 3.5 4.75 7.75" />
-			</svg>
+			<ArrowUpIcon aria-hidden="true" className="send-action-icon" size={18} />
 		</button>
 	);
 }
